Move utils functions from App.js
